refactor : add default parameters

+++ app/src/main/java/com/android/sample/ui/calendar/utils/WeekViewMetrics.kt
@@ -15,7 +15,7 @@ internal data class WeekViewMetrics(
     val effectiveEndTime: LocalTime,
     val gridStartTime: LocalTime,
     val rowHeightDp: Dp,
-    val totalHours: Float,
+    val totalHours: Int,
     val gridHeightDp: Dp,
     val timeLabels: List<LocalTime>,
     val visibleTimeSpan: TimeSpan,
